+++ ReleaseNotes.md
@@ -1,7 +1,13 @@
 # Release Notes
 > **v3.0.0**
-> - Added test coverage.
+> - Added new logo.
+> - Changed 'PackageLicenseUrl' element to 'PackageLicenseFile'.
+> - Changed 'PackageIconUrl' element to 'PackageIcon'.
+> - The build will now fail when no source files are found.
+> - Started to add unit test coverage.
 > - Added output variables to task GUI.
+> - Added support for `MSBuild.Sdk.Extras` projects.
+> Thanks to [crb04c](https://github.com/crb04c)
 
 > **v2.3.105**
 > - Added support for `.vbproj` to .Net Core task.